+++ iOS/Frameworks/MessageUI.framework/MFComposeTypeFactory.h
@@ -10,7 +10,7 @@
 {
 }
 
-+ (id)subjectFromSubject:(id)arg1 withComposeType:(int)arg2;
++ (id)subjectFromSubject:(id)arg1 withComposeType:(long long)arg2;
 + (void)hijackThreadFromDelegate:(id)arg1;
 + (id)messageFromDelegate:(id)arg1 originatingBundleID:(id)arg2 sourceAccountManagement:(int)arg3;
 + (id)messageFromDelegate:(id)arg1 withSubstituteDOMDocument:(id)arg2 compositionSpecification:(id)arg3 originatingBundleID:(id)arg4 sourceAccountManagement:(int)arg5 writeAttachmentPlaceholders:(_Bool)arg6;
