+++ iOS/Frameworks/NetworkExtension.framework/NEVPNPluginManager-Protocol.h
@@ -8,7 +8,7 @@
 
 @protocol NEVPNPluginManager
 - (void)setTunnelNetworkSettings:(NETunnelNetworkSettings *)arg1 completionHandler:(void (^)(NSError *))arg2;
-- (void)setStatus:(long long)arg1 error:(NSError *)arg2;
+- (void)setStatus:(int)arg1 error:(NSError *)arg2;
 - (void)handleIPCDetached;
 - (void)createVirtualInterfaceOfType:(long long)arg1 maxPendingPackets:(unsigned long long)arg2 completionHandler:(void (^)(NSFileHandle *))arg3;
 @end
