+++ iOS/PrivateFrameworks/SpringBoard.framework/SBDeviceApplicationSceneStatusBarStateObserver-Protocol.h
@@ -14,7 +14,7 @@
 - (void)didInvalidateStatusBarDescriptionForSceneWithIdentifier:(NSString *)arg1;
 - (void)sceneWithIdentifier:(NSString *)arg1 didChangeSceneInterfaceOrientationTo:(long long)arg2;
 - (void)sceneWithIdentifier:(NSString *)arg1 didChangeStatusBarAvoidanceFrameTo:(struct CGRect)arg2;
-- (void)sceneWithIdentifier:(NSString *)arg1 didChangeStatusBarStyleOverridesToSuppressTo:(int)arg2;
+- (void)sceneWithIdentifier:(NSString *)arg1 didChangeStatusBarStyleOverridesToSuppressTo:(unsigned long long)arg2;
 - (void)sceneWithIdentifier:(NSString *)arg1 didChangeStatusBarOrientationTo:(long long)arg2;
 - (void)sceneWithIdentifier:(NSString *)arg1 didChangeStatusBarHiddenTo:(_Bool)arg2 withAnimation:(long long)arg3;
 - (void)sceneWithIdentifier:(NSString *)arg1 didChangeStatusBarAlphaTo:(double)arg2;
