+++ iOS/PrivateFrameworks/SpringBoard.framework/SBDeviceApplicationSceneStatusBarStateProvider_Base.h
@@ -40,7 +40,7 @@
 @property(readonly, nonatomic) _UIStatusBarData *overlayStatusBarData;
 @property(readonly, nonatomic) _Bool sceneWantsDeviceOrientationEventsEnabled;
 @property(readonly, nonatomic) struct CGRect statusBarAvoidanceFrame;
-@property(readonly, nonatomic) int statusBarStyleOverridesToSuppress;
+@property(readonly, nonatomic) unsigned long long statusBarStyleOverridesToSuppress;
 @property(readonly, nonatomic) long long statusBarOrientation;
 @property(readonly, nonatomic) _Bool statusBarHidden;
 @property(readonly, nonatomic) double statusBarAlpha;
