+++ iOS/PrivateFrameworks/SpringBoard.framework/SBStatusBarTapManager.h
@@ -20,15 +20,15 @@
 - (id)_allApplicationDestinationsForApplications:(id)arg1;
 - (_Bool)_isUILocked;
 - (id)_workspace;
-- (void)_resetRecentlyTappedSceneIdentifiersForStyleOverrides:(int)arg1;
-- (id)_recentlyTappedApplicationDestinationsForStyleOverrides:(int)arg1;
-- (void)_scheduleResetRecentlyTappedSceneIdentifiersForStyleOverrides:(int)arg1;
-- (void)_addRecentlyTappedApplicationDestinations:(id)arg1 forStyleOverrides:(int)arg2;
-- (id)_appDestinationForNextSceneToVisitForApps:(id)arg1 styleOverrides:(int)arg2;
-- (id)_handlerForStyleOverrides:(int)arg1 orOutApplications:(id *)arg2;
-- (id)_handlerForStyleOverrides:(int)arg1;
-- (_Bool)canHandleTapForStatusBarStyleOverrides:(int)arg1;
-- (_Bool)handleTapForStyleOverrides:(int)arg1;
+- (void)_resetRecentlyTappedSceneIdentifiersForStyleOverrides:(unsigned long long)arg1;
+- (id)_recentlyTappedApplicationDestinationsForStyleOverrides:(unsigned long long)arg1;
+- (void)_scheduleResetRecentlyTappedSceneIdentifiersForStyleOverrides:(unsigned long long)arg1;
+- (void)_addRecentlyTappedApplicationDestinations:(id)arg1 forStyleOverrides:(unsigned long long)arg2;
+- (id)_appDestinationForNextSceneToVisitForApps:(id)arg1 styleOverrides:(unsigned long long)arg2;
+- (id)_handlerForStyleOverrides:(unsigned long long)arg1 orOutApplications:(id *)arg2;
+- (id)_handlerForStyleOverrides:(unsigned long long)arg1;
+- (_Bool)canHandleTapForStatusBarStyleOverrides:(unsigned long long)arg1;
+- (_Bool)handleTapForStyleOverrides:(unsigned long long)arg1;
 
 @end
