+++ iOS/PrivateFrameworks/SpringBoard.framework/UIStatusBarStyleDelegate_SpringBoardOnly-Protocol.h
@@ -13,6 +13,6 @@
 @optional
 - (_Bool)statusBarShouldDisableRegionActions:(UIStatusBar_Base *)arg1;
 - (long long)overriddenRequestedStyleFromStyle:(long long)arg1;
-- (int)statusBar:(UIStatusBar_Base *)arg1 effectiveStyleOverridesForRequestedStyle:(long long)arg2 overrides:(int)arg3;
+- (unsigned long long)statusBar:(UIStatusBar_Base *)arg1 effectiveStyleOverridesForRequestedStyle:(long long)arg2 overrides:(unsigned long long)arg3;
 @end
