+++ iOS/PrivateFrameworks/SpringBoardServices.framework/SBSStatusBarStyleOverridesCoordinator.h
@@ -10,16 +10,16 @@
 
 @interface SBSStatusBarStyleOverridesCoordinator : NSObject
 {
-    int _styleOverrides;
+    unsigned long long _styleOverrides;
     id <SBSStatusBarStyleOverridesCoordinatorDelegate> _delegate;
 }
 
 - (void).cxx_destruct;
-@property(readonly, nonatomic) int styleOverrides; // @synthesize styleOverrides=_styleOverrides;
+@property(readonly, nonatomic) unsigned long long styleOverrides; // @synthesize styleOverrides=_styleOverrides;
 @property(nonatomic) __weak id <SBSStatusBarStyleOverridesCoordinatorDelegate> delegate; // @synthesize delegate=_delegate;
 - (void)_registrationInvalidated:(id)arg1;
 - (void)_handleStatusBarTapWithContext:(id)arg1 withCompletionBlock:(CDUnknownBlockType)arg2;
-- (void)setRegisteredStyleOverrides:(int)arg1 reply:(CDUnknownBlockType)arg2;
+- (void)setRegisteredStyleOverrides:(unsigned long long)arg1 reply:(CDUnknownBlockType)arg2;
 - (void)dealloc;
 
 @end
